requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
amadeus>=9.0.0
googlemaps>=4.10.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dateutil>=2.8.0
